requirements.txt
h5py>=2.10.0
numpy>=1.18.1
tqdm>=4.43.0
albumentations==1.1.0
imgaug==0.4.0
matplotlib==3.5.1
opencv-python-headless==4.5.5.62
pandas==1.3.5
Pillow==9.0.0
scikit-image==0.19.1
scikit-learn==1.0.2
scipy==1.7.3
timm==0.4.12
torch==1.8.0
torch-optimizer==0.3.0
torchvision==0.9.0
wandb==0.12.9
